Allow to reuse required services in fido2 server
