+++ Server/src/main/java/org/gluu/oxauth/auth/Authenticator.java
@@ -24,7 +24,12 @@ import org.gluu.oxauth.model.jwt.JwtClaimName;
 import org.gluu.oxauth.model.registration.Client;
 import org.gluu.oxauth.model.util.Util;
 import org.gluu.oxauth.security.Identity;
-import org.gluu.oxauth.service.*;
+import org.gluu.oxauth.service.AuthenticationService;
+import org.gluu.oxauth.service.ClientService;
+import org.gluu.oxauth.service.ErrorHandlerService;
+import org.gluu.oxauth.service.RequestParameterService;
+import org.gluu.oxauth.service.SessionIdService;
+import org.gluu.oxauth.service.common.*;
 import org.gluu.oxauth.service.external.ExternalAuthenticationService;
 import org.gluu.util.Pair;
 import org.gluu.util.StringHelper;
