+++ Server/src/main/java/org/gluu/oxauth/authorize/ws/rs/AuthorizeAction.java
@@ -30,7 +30,15 @@ import org.gluu.oxauth.model.registration.Client;
 import org.gluu.oxauth.model.util.Base64Util;
 import org.gluu.oxauth.model.util.JwtUtil;
 import org.gluu.oxauth.model.util.Util;
-import org.gluu.oxauth.service.*;
+import org.gluu.oxauth.service.AuthorizeService;
+import org.gluu.oxauth.service.ClientAuthorizationsService;
+import org.gluu.oxauth.service.ClientService;
+import org.gluu.oxauth.service.CookieService;
+import org.gluu.oxauth.service.ErrorHandlerService;
+import org.gluu.oxauth.service.RedirectionUriService;
+import org.gluu.oxauth.service.RequestParameterService;
+import org.gluu.oxauth.service.SessionIdService;
+import org.gluu.oxauth.service.common.*;
 import org.gluu.oxauth.service.external.ExternalAuthenticationService;
 import org.gluu.oxauth.service.external.ExternalConsentGatheringService;
 import org.gluu.oxauth.service.external.ExternalPostAuthnService;
