+++ Server/src/main/java/org/gluu/oxauth/authorize/ws/rs/AuthorizeRestWebServiceImpl.java
@@ -37,7 +37,15 @@ import org.gluu.oxauth.model.util.Base64Util;
 import org.gluu.oxauth.model.util.JwtUtil;
 import org.gluu.oxauth.model.util.Util;
 import org.gluu.oxauth.security.Identity;
-import org.gluu.oxauth.service.*;
+import org.gluu.oxauth.service.AuthenticationFilterService;
+import org.gluu.oxauth.service.ClientAuthorizationsService;
+import org.gluu.oxauth.service.ClientService;
+import org.gluu.oxauth.service.CookieService;
+import org.gluu.oxauth.service.RedirectUriResponse;
+import org.gluu.oxauth.service.RequestParameterService;
+import org.gluu.oxauth.service.SessionIdService;
+import org.gluu.oxauth.service.UserService;
+import org.gluu.oxauth.service.common.*;
 import org.gluu.oxauth.service.external.ExternalPostAuthnService;
 import org.gluu.oxauth.service.external.context.ExternalPostAuthnContext;
 import org.gluu.oxauth.util.QueryStringDecoder;
