+++ Server/src/main/java/org/gluu/oxauth/model/common/AuthorizationGrant.java
@@ -22,7 +22,11 @@ import org.gluu.oxauth.model.token.IdTokenFactory;
 import org.gluu.oxauth.model.token.JsonWebResponse;
 import org.gluu.oxauth.model.token.JwtSigner;
 import org.gluu.oxauth.model.util.JwtUtil;
-import org.gluu.oxauth.service.*;
+import org.gluu.oxauth.service.AttributeService;
+import org.gluu.oxauth.service.ClientService;
+import org.gluu.oxauth.service.GrantService;
+import org.gluu.oxauth.service.SectorIdentifierService;
+import org.gluu.oxauth.service.common.*;
 import org.gluu.oxauth.service.external.ExternalIntrospectionService;
 import org.gluu.oxauth.service.external.context.ExternalIntrospectionContext;
 import org.gluu.oxauth.util.TokenHashUtil;
