+++ Server/src/main/java/org/gluu/oxauth/register/ws/rs/RegisterRestWebServiceImpl.java
@@ -38,7 +38,11 @@ import org.gluu.oxauth.model.token.HandleTokenFactory;
 import org.gluu.oxauth.model.util.JwtUtil;
 import org.gluu.oxauth.model.util.Pair;
 import org.gluu.oxauth.model.util.Util;
-import org.gluu.oxauth.service.*;
+import org.gluu.oxauth.service.AttributeService;
+import org.gluu.oxauth.service.ClientService;
+import org.gluu.oxauth.service.MetricService;
+import org.gluu.oxauth.service.ScopeService;
+import org.gluu.oxauth.service.common.*;
 import org.gluu.oxauth.service.external.ExternalDynamicClientRegistrationService;
 import org.gluu.oxauth.service.token.TokenService;
 import org.gluu.oxauth.util.ServerUtil;
