+++ Server/src/main/java/org/gluu/oxauth/service/LdapCustomAuthenticationConfigurationService.java
@@ -17,6 +17,7 @@ import javax.inject.Named;
 import org.gluu.model.AuthenticationScriptUsageType;
 import org.gluu.model.SimpleCustomProperty;
 import org.gluu.model.config.CustomAuthenticationConfiguration;
+import org.gluu.oxauth.service.common.ConfigurationService;
 import org.gluu.util.StringHelper;
 import org.oxauth.persistence.model.configuration.CustomProperty;
 import org.oxauth.persistence.model.configuration.GluuConfiguration;
