+++ Server/src/main/java/org/gluu/oxauth/service/MetricService.java
@@ -8,6 +8,8 @@ package org.gluu.oxauth.service;
 
 import org.gluu.model.ApplicationType;
 import org.gluu.oxauth.model.configuration.AppConfiguration;
+import org.gluu.oxauth.service.common.ApplicationFactory;
+import org.gluu.oxauth.service.common.ConfigurationService;
 import org.gluu.persist.PersistenceEntryManager;
 import org.gluu.service.metric.inject.ReportMetric;
 import org.gluu.oxauth.model.config.StaticConfiguration;
