+++ Server/src/main/java/org/gluu/oxauth/service/external/ExternalAuthenticationService.java
@@ -30,8 +30,8 @@ import org.gluu.model.custom.script.model.auth.AuthenticationCustomScript;
 import org.gluu.model.custom.script.type.BaseExternalType;
 import org.gluu.model.custom.script.type.auth.PersonAuthenticationType;
 import org.gluu.model.ldap.GluuLdapConfiguration;
-import org.gluu.oxauth.service.ApplicationFactory;
 import org.gluu.oxauth.service.cdi.event.ReloadAuthScript;
+import org.gluu.oxauth.service.common.ApplicationFactory;
 import org.gluu.oxauth.service.external.internal.InternalDefaultPersonAuthenticationType;
 import org.gluu.service.custom.script.ExternalScriptService;
 import org.gluu.util.OxConstants;
