+++ Server/src/main/java/org/gluu/oxauth/service/external/context/ExternalResourceOwnerPasswordCredentialsContext.java
@@ -3,8 +3,8 @@ package org.gluu.oxauth.service.external.context;
 import org.gluu.model.custom.script.conf.CustomScriptConfiguration;
 import org.gluu.oxauth.model.configuration.AppConfiguration;
 import org.gluu.oxauth.service.AttributeService;
+import org.gluu.oxauth.service.common.UserService;
 import org.gluu.oxauth.model.common.User;
-import org.gluu.oxauth.service.UserService;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
