+++ Server/src/main/java/org/gluu/oxauth/service/fido/u2f/AuthenticationService.java
@@ -35,12 +35,12 @@ import org.gluu.oxauth.model.fido.u2f.protocol.AuthenticateRequestMessage;
 import org.gluu.oxauth.model.fido.u2f.protocol.AuthenticateResponse;
 import org.gluu.oxauth.model.fido.u2f.protocol.ClientData;
 import org.gluu.oxauth.model.util.Base64Util;
+import org.gluu.oxauth.service.common.UserService;
 import org.gluu.persist.PersistenceEntryManager;
 import org.gluu.search.filter.Filter;
 import org.gluu.util.StringHelper;
 import org.slf4j.Logger;
 import org.gluu.oxauth.model.config.StaticConfiguration;
-import org.gluu.oxauth.service.UserService;
 
 /**
  * Provides operations with U2F authentication request
