+++ Server/src/main/java/org/gluu/oxauth/service/fido/u2f/DeviceRegistrationService.java
@@ -9,6 +9,7 @@ package org.gluu.oxauth.service.fido.u2f;
 import org.gluu.oxauth.model.fido.u2f.DeviceRegistration;
 import org.gluu.oxauth.model.fido.u2f.DeviceRegistrationStatus;
 import org.gluu.oxauth.model.util.Base64Util;
+import org.gluu.oxauth.service.common.UserService;
 import org.gluu.persist.PersistenceEntryManager;
 import org.gluu.persist.model.BatchOperation;
 import org.gluu.persist.model.SearchScope;
@@ -17,7 +18,6 @@ import org.gluu.search.filter.Filter;
 import org.gluu.util.StringHelper;
 import org.slf4j.Logger;
 import org.gluu.oxauth.model.config.StaticConfiguration;
-import org.gluu.oxauth.service.UserService;
 
 import javax.ejb.Stateless;
 import javax.inject.Inject;
