+++ Server/src/main/java/org/gluu/oxauth/service/fido/u2f/ValidationService.java
@@ -14,10 +14,10 @@ import org.gluu.oxauth.model.common.SessionId;
 import org.gluu.oxauth.model.config.Constants;
 import org.gluu.oxauth.model.fido.u2f.U2fConstants;
 import org.gluu.oxauth.service.SessionIdService;
+import org.gluu.oxauth.service.common.UserService;
 import org.gluu.util.StringHelper;
 import org.slf4j.Logger;
 import org.gluu.oxauth.model.common.User;
-import org.gluu.oxauth.service.UserService;
 
 /**
  * Utility to validate U2F input data
