+++ Server/src/main/java/org/gluu/oxauth/service/push/sns/PushSnsService.java
@@ -10,7 +10,7 @@ import javax.inject.Inject;
 import javax.inject.Named;
 
 import org.gluu.oxauth.model.configuration.AppConfiguration;
-import org.gluu.oxauth.service.EncryptionService;
+import org.gluu.oxauth.service.common.EncryptionService;
 import org.gluu.oxauth.util.ServerUtil;
 import org.gluu.persist.PersistenceEntryManager;
 import org.gluu.oxauth.model.common.User;
