+++ Server/src/main/java/org/gluu/oxauth/session/ws/rs/EndSessionRestWebServiceImpl.java
@@ -27,7 +27,12 @@ import org.gluu.oxauth.model.session.EndSessionErrorResponseType;
 import org.gluu.oxauth.model.token.JsonWebResponse;
 import org.gluu.oxauth.model.util.URLPatternList;
 import org.gluu.oxauth.model.util.Util;
-import org.gluu.oxauth.service.*;
+import org.gluu.oxauth.service.ClientService;
+import org.gluu.oxauth.service.CookieService;
+import org.gluu.oxauth.service.GrantService;
+import org.gluu.oxauth.service.RedirectionUriService;
+import org.gluu.oxauth.service.SessionIdService;
+import org.gluu.oxauth.service.common.*;
 import org.gluu.oxauth.service.external.ExternalApplicationSessionService;
 import org.gluu.oxauth.service.external.ExternalEndSessionService;
 import org.gluu.oxauth.service.external.context.EndSessionContext;
