+++ Server/src/main/java/org/gluu/oxauth/token/ws/rs/TokenRestWebServiceImpl.java
@@ -25,7 +25,13 @@ import org.gluu.oxauth.model.token.JwrService;
 import org.gluu.oxauth.model.token.TokenErrorResponseType;
 import org.gluu.oxauth.model.token.TokenParamsValidator;
 import org.gluu.oxauth.security.Identity;
-import org.gluu.oxauth.service.*;
+import org.gluu.oxauth.service.AttributeService;
+import org.gluu.oxauth.service.AuthenticationFilterService;
+import org.gluu.oxauth.service.AuthenticationService;
+import org.gluu.oxauth.service.GrantService;
+import org.gluu.oxauth.service.SessionIdService;
+import org.gluu.oxauth.service.UserService;
+import org.gluu.oxauth.service.common.*;
 import org.gluu.oxauth.service.external.ExternalResourceOwnerPasswordCredentialsService;
 import org.gluu.oxauth.service.external.context.ExternalResourceOwnerPasswordCredentialsContext;
 import org.gluu.oxauth.uma.service.UmaTokenService;
