+++ Server/src/main/java/org/gluu/oxauth/uma/service/UmaScopeService.java
@@ -13,8 +13,8 @@ import org.gluu.oxauth.model.configuration.AppConfiguration;
 import org.gluu.oxauth.model.error.ErrorResponseFactory;
 import org.gluu.oxauth.model.registration.Client;
 import org.gluu.oxauth.model.uma.UmaErrorResponseType;
-import org.gluu.oxauth.service.InumService;
 import org.gluu.oxauth.service.SpontaneousScopeService;
+import org.gluu.oxauth.service.common.InumService;
 import org.gluu.persist.PersistenceEntryManager;
 import org.gluu.search.filter.Filter;
 import org.oxauth.persistence.model.Scope;
