+++ Server/src/main/java/org/gluu/oxauth/userinfo/ws/rs/UserInfoRestWebServiceImpl.java
@@ -41,7 +41,12 @@ import org.gluu.oxauth.model.userinfo.UserInfoErrorResponseType;
 import org.gluu.oxauth.model.userinfo.UserInfoParamsValidator;
 import org.gluu.oxauth.model.util.JwtUtil;
 import org.gluu.oxauth.model.util.Util;
-import org.gluu.oxauth.service.*;
+import org.gluu.oxauth.service.AttributeService;
+import org.gluu.oxauth.service.ClientService;
+import org.gluu.oxauth.service.ScopeService;
+import org.gluu.oxauth.service.ServerCryptoProvider;
+import org.gluu.oxauth.service.UserService;
+import org.gluu.oxauth.service.common.*;
 import org.gluu.oxauth.service.external.ExternalDynamicScopeService;
 import org.gluu.oxauth.service.external.context.DynamicScopeExternalContext;
 import org.gluu.oxauth.service.token.TokenService;
