+++ Server/src/main/java/org/gluu/oxauth/ws/rs/fido/u2f/U2fAuthenticationWS.java
@@ -21,7 +21,7 @@ import org.gluu.oxauth.model.fido.u2f.protocol.AuthenticateRequestMessage;
 import org.gluu.oxauth.model.fido.u2f.protocol.AuthenticateResponse;
 import org.gluu.oxauth.model.fido.u2f.protocol.AuthenticateStatus;
 import org.gluu.oxauth.model.util.Base64Util;
-import org.gluu.oxauth.service.UserService;
+import org.gluu.oxauth.service.common.UserService;
 import org.gluu.oxauth.service.fido.u2f.AuthenticationService;
 import org.gluu.oxauth.service.fido.u2f.DeviceRegistrationService;
 import org.gluu.oxauth.service.fido.u2f.UserSessionIdService;
