+++ Server/src/test/java/org/gluu/oxauth/comp/KeyGenerationTest.java
@@ -9,7 +9,7 @@ package org.gluu.oxauth.comp;
 import javax.inject.Inject;
 
 import org.gluu.oxauth.BaseTest;
-import org.gluu.oxauth.service.EncryptionService;
+import org.gluu.oxauth.service.common.EncryptionService;
 import org.testng.annotations.Parameters;
 import org.testng.annotations.Test;
