+++ Server/src/test/java/org/gluu/oxauth/comp/SessionIdServiceTest.java
@@ -21,10 +21,10 @@ import org.gluu.oxauth.BaseComponentTest;
 import org.gluu.oxauth.model.common.SessionId;
 import org.gluu.oxauth.model.common.SessionIdState;
 import org.gluu.oxauth.service.SessionIdService;
+import org.gluu.oxauth.service.common.UserService;
 import org.testng.Assert;
 import org.testng.annotations.Parameters;
 import org.testng.annotations.Test;
-import org.gluu.oxauth.service.UserService;
 
 /**
  * @author Yuriy Zabrovarnyy
