+++ Server/src/test/java/org/gluu/oxauth/comp/UmaResourceServiceTest.java
@@ -6,7 +6,7 @@ import org.gluu.oxauth.model.registration.Client;
 import org.gluu.oxauth.model.token.HandleTokenFactory;
 import org.gluu.oxauth.model.uma.persistence.UmaResource;
 import org.gluu.oxauth.service.ClientService;
-import org.gluu.oxauth.service.InumService;
+import org.gluu.oxauth.service.common.InumService;
 import org.gluu.oxauth.uma.service.UmaResourceService;
 import org.gluu.util.security.StringEncrypter;
 import org.testng.annotations.Test;
