+++ common/src/main/java/org/gluu/oxauth/service/common/ConfigurationService.java
@@ -4,7 +4,7 @@
  * Copyright (c) 2014, Gluu
  */
 
-package org.gluu.oxauth.service;
+package org.gluu.oxauth.service.common;
 
 import javax.enterprise.context.ApplicationScoped;
 import javax.inject.Inject;
